pyproject.toml
[project]
name = "erp"
version = "0.1.0"
description = ""
authors = [
    {name = "Jace Trent",email = "trentjace58@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "environs (>=14.5.0,<15.0.0)",
    "django-unfold (>=0.76.0,<0.77.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

